pyproject.toml
[tool.poetry]
name = "agenticscraping"
version = "0.1.0"
description = "Agentic scraping framework"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
uvicorn = "*"
celery = "*"
crewai = "*"
playwright = "*"
sqlalchemy = "*"
psycopg = {extras = ["binary"], version="*"}
httpx = "*"
beautifulsoup4 = "*"
opentelemetry-api = "*"
opentelemetry-sdk = "*"
opentelemetry-instrumentation-fastapi = "*"
opentelemetry-instrumentation-celery = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"

